REFACTORIZED some files of the UI interface
